微修正,テスト修正

+++ db/migrate/20200124093201_change_data_valuation_to_points.rb
@@ -1,5 +1,5 @@
 class ChangeDataValuationToPoints < ActiveRecord::Migration[6.0]
   def change
-    change_column :points, :valuation, :float
+    change_column :points, :valuation, :integer
   end
 end
